requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.5
django-auditlog==3.0.0
django-background-tasks==1.2.8
django-cors-headers==4.5.0
django-crispy-forms==2.0
django-filter==23.5
django-pandas==0.6.7
djangorestframework==3.15.1
djangorestframework-api-key==3.0.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
kombu==5.5.4
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
